Make two separate category responses
