+++ data-workflows/plugin/tests/test_metadata.py
@@ -5,7 +5,6 @@ import pytest
 
 from plugin import metadata, categories
 from plugin.metadata import get_formatted_metadata
-from plugin.tests.utils import category_responses
 
 PLUGIN = "napari-demo"
 VERSION = "0.0.2"
@@ -18,6 +17,19 @@ ONTOLOGY = "EDAM-BIOIMAGING:alpha06"
 def category_terms():
     return ["Img reg", "2D image", "Affine reg", "Foo"]
 
+def category_responses():
+    return [
+        [{"dimension": "Wrkflw", "hierarchy": ["Img reg"], "label": "Img reg1"}],
+        [{"dimension": "Data", "hierarchy": ["2D image"], "label": "2D"}],
+        [
+            {
+                "dimension": "Wrkflw",
+                "hierarchy": ["Img reg", "Affine reg"],
+                "label": "Img reg2",
+            }
+        ],
+        [],
+    ]
 
 class TestMetadata:
     @pytest.fixture
