--- data-workflows/plugin/tests/utils.py
@@ -1,13 +1,0 @@
-def category_responses():
-    return [
-        [{"dimension": "Wrkflw", "hierarchy": ["Img reg"], "label": "Img reg1"}],
-        [{"dimension": "Data", "hierarchy": ["2D image"], "label": "2D"}],
-        [
-            {
-                "dimension": "Wrkflw",
-                "hierarchy": ["Img reg", "Affine reg"],
-                "label": "Img reg2",
-            }
-        ],
-        [],
-    ]
